Remove jasper option from ImageMagick6
Remove the optional jasper option to link against graphics/jasper. With
the default options this causes ImageMagick to pull in graphics/jasper
at install. However ImageMagick does not actually link against
graphics/jasper and has not done since < 6.8 versions.

jasper was linked in to support the jp2 format, which it now supports
via linking against openjpeg.

So we add the option 'jp2' which links against openjpeg. We also remove
it from the suggested options as it's not a common format and hasn't
actually been supported in pkgsrc ImageMagick for some time.

We also bump the latest minor patch release while in here.